chore: refactor of new ifc handling store class
