+++ src/ada/ifc/write/write_plates.py
@@ -70,6 +70,6 @@ def write_ifc_plate(plate: Plate):
         add_colour(f, ifcextrudedareasolid, str(plate.colour), plate.colour)
 
     # Material
-    ifc_store.associate_elem_with_material(plate.material, ifc_plate)
+    ifc_store.writer.associate_elem_with_material(plate.material, ifc_plate)
 
     return ifc_plate
